initiated todomvc example, add viewOnBoth, replaced CRC64 with hasCode()

+++ vertxui-core/src/main/java/live/connector/vertxui/client/fluent/Viewable.java
@@ -8,6 +8,6 @@ package live.connector.vertxui.client.fluent;
  */
 public interface Viewable {
 
-	public long getCrc();
+	public int getCrc();
 
 }
